+++ src/main/java/com/ireyes/findMyPet/model/user/register/OnValidationTokenExpire.java
@@ -22,7 +22,9 @@ public class OnValidationTokenExpire implements Runnable{
 	
 	@Override
 	public void run() {
-		repo.deleteById(id);
+		if(repo.existsById(id)) {
+			repo.deleteById(id);
+		}
 		logger.info(() -> "Expired ValidationToken with id " + id.longValue());
 	}
